+++ src/netprotocols/layer3/__init__.py
@@ -1,21 +1,28 @@
-from netprotocols.layer3.icmp import ICMPv4, ICMPv6
-from netprotocols.layer3.igmp import IGMP
-from netprotocols.layer3.ip import IPv4, IPv6
+from netprotocols.layer3.gre import GRE
+from netprotocols.layer3.icmp import ICMPv4, ICMPv6, NDPOption
+from netprotocols.layer3.igmp import IGMP, IGMPv3GroupRecord
+from netprotocols.layer3.ip import IPv4, IPv4Option, IPv6
 from netprotocols.layer3.ipv6_ext import (
     IPv6DestinationOptions,
     IPv6Fragment,
     IPv6HopByHopOptions,
+    IPv6Option,
     IPv6Routing,
 )
 
 __all__ = [
+    "GRE",
     "IGMP",
     "ICMPv4",
     "ICMPv6",
+    "IGMPv3GroupRecord",
     "IPv4",
+    "IPv4Option",
     "IPv6",
     "IPv6DestinationOptions",
     "IPv6Fragment",
     "IPv6HopByHopOptions",
+    "IPv6Option",
     "IPv6Routing",
+    "NDPOption",
 ]