+++ src/netprotocols/layer4/__init__.py
@@ -1,4 +1,4 @@
-from netprotocols.layer4.tcp import TCP
+from netprotocols.layer4.tcp import TCP, TCPOption
 from netprotocols.layer4.udp import UDP
 
-__all__ = ["TCP", "UDP"]
+__all__ = ["TCP", "UDP", "TCPOption"]